% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_and_resample.R
\name{download_and_resample}
\alias{download_and_resample}
\title{Get data from province of Bozen monitoring stations. applies a resample with "resample_provBz_data"}
\usage{
download_and_resample(sensor, station, datestart, dateend,
  round = "hour", spread = FALSE, notScode = FALSE,
  inshiny = FALSE, nstations = NULL, sort = NULL)
}
\arguments{
\item{sensor}{Sensor to download. You can check SCODES station names and sensors typing MeteoBrowser::tot_tab_def}

\item{station}{Station "SCODE" to download. You can check SCODES station names and sensors typing MeteoBrowser::tot_tab_def}

\item{datestart}{starting date of the timeseries. format "YYYY-MM-DD"}

\item{dateend}{ending date of the timeseries. format "YYYY-MM-DD"}

\item{round}{the timestamp of the resample. defalutl is "hour" . write "raw" for no resample}

\item{spread}{whether to spread the table or leave it in long format. Default is FALSE}

\item{inshiny}{if the function is called in the shiny app. default=FALSE}

\item{nstations}{if the function is called in the shiny app: number of stations * sensors. default=NULL}

\item{sort}{column names on which to sort the table. available options: c("NAME","TimeStamp","Sensor","Value")}
}
\description{
Get data from province of Bozen monitoring stations. applies a resample with "resample_provBz_data"
}
