% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_and_resample2.R
\name{download_and_resample2}
\alias{download_and_resample2}
\title{Get data from province of Bozen monitoring stations. applies a resample with "resample_provBz_data2"}
\usage{
download_and_resample2(sensor, station, datestart, dateend, roundFUN,
  round = "hour", statsFUN, spread = FALSE, notScode = FALSE,
  inshiny = FALSE, nstations = NULL, sort = NULL, name = NULL, ...)
}
\arguments{
\item{sensor}{Sensor to download. You can check SCODES station names and sensors typing MeteoBrowser::tot_tab_def}

\item{station}{Station "SCODE" to download. You can check SCODES station names and sensors typing MeteoBrowser::tot_tab_def}

\item{datestart}{starting date of the timeseries. format "YYYY-MM-DD"}

\item{dateend}{ending date of the timeseries. format "YYYY-MM-DD"}

\item{roundFUN}{the resampling function. Possible options lubridate::ceiling_date,lubridate::floor_date,lubridate::round_date}

\item{round}{the resampling period. Possible options "raw" "hour" "day" "week" "month". Passed to lubridate::round_date}

\item{statsFUN}{the "statistical" function. any function that takes a vector in inpunt and a single value in output}

\item{spread}{whether to spread the table or leave it in long format. Default is FALSE}

\item{inshiny}{if the function is called in the shiny app. default=FALSE}

\item{nstations}{if the function is called in the shiny app: number of stations * sensors. default=NULL}

\item{sort}{column names on which to sort the table. available options: "NAME","TimeStamp","Sensor","Value"}

\item{name}{a name to give to a column "NAME". default is the german/italian name of the station}

\item{...}{additional arguments passed to statsFUN}
}
\description{
Get data from province of Bozen monitoring stations. applies a resample with "resample_provBz_data2"
}
