% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sensor.R
\name{download_sensor}
\alias{download_sensor}
\title{internal function Get data one sensor-sation at a time from province of Bozen monitoring stations.}
\usage{
download_sensor(sensor, station, datestart, dateend, round = "hour",
  notScode = FALSE, inshiny = FALSE, nstations = NULL)
}
\arguments{
\item{sensor}{Sensor code}

\item{station}{SCODE of the station}

\item{datestart}{starting date of the timeseries.}

\item{dateend}{ending date of the timeseries.}

\item{round}{the timestamp of the resample. defalutl is "hour" . write "raw" for no resample}

\item{notScode}{if TRUE SCODE column is deleted}

\item{spread}{whether to spread the table or leave it in long format}
}
\description{
internal function Get data one sensor-sation at a time from province of Bozen monitoring stations.
}
