% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_station.R
\name{download_station}
\alias{download_station}
\title{internal function Get data form one sation and multiple sensors from province of Bozen monitoring stations.}
\usage{
download_station(station, station_sensor, datestart, dateend,
  sensors = unique(get_provBz_sensors()$Sensor), round = "hour",
  spread = FALSE, notScode = FALSE, inshiny = FALSE,
  nstations = NULL)
}
\arguments{
\item{station}{SCODE of the station}

\item{datestart}{starting date of the timeseries.}

\item{dateend}{ending date of the timeseries.}

\item{round}{the timestamp of the resample. defalutl is "hour" . write "raw" for no resample}

\item{spread}{whether to spread the table or leave it in long format}

\item{notScode}{if TRUE SCODE column is deleted}

\item{sensor}{Sensor code}
}
\description{
internal function Get data form one sation and multiple sensors from province of Bozen monitoring stations.
}
