% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provBz_data.R
\name{get_provBz_data}
\alias{get_provBz_data}
\title{Get data from province of Bozen monitoring stations.}
\usage{
get_provBz_data(station_sensor, datestart = Sys.Date() - 1,
  dateend = Sys.Date() + 1, round = "hour", spread = FALSE,
  nstations = NULL, notScode = FALSE)
}
\arguments{
\item{station_sensor}{a table with SCODE and Sensor columns.}

\item{datestart}{starting date of the timeseries.}

\item{dateend}{ending date of the timeseries.}

\item{round}{the timestamp of the resample. defalutl is "hour" . write "raw" for no resample}

\item{spread}{whether to spread the table or leave it in long format}
}
\description{
Get data from province of Bozen monitoring stations.
}
