% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provBz_sensors.R
\name{get_provBz_sensors}
\alias{get_provBz_sensors}
\title{Get data from province of Bozen sensors. Internal function}
\usage{
get_provBz_sensors()
}
\description{
Get data from province of Bozen sensors. Internal function
}
