% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeteoLeaflet.R
\name{plotMeteoLeaflet}
\alias{plotMeteoLeaflet}
\title{display a leaflet map with all the metereological stations of the province of bozen}
\usage{
plotMeteoLeaflet()
}
\description{
display a leaflet map with all the metereological stations of the province of bozen
}
