% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass_wind.R
\name{reclass_wind}
\alias{reclass_wind}
\title{reclassify wind from degrees to 8 directions}
\usage{
reclass_wind(wind)
}
\arguments{
\item{wind}{a vector with values between 0 and 360}
}
\description{
reclassify wind from degrees to 8 directions
}
