% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_provBz_data.R
\name{resample_provBz_data}
\alias{resample_provBz_data}
\title{resamples data from the province of Bozen monitoring stations.}
\usage{
resample_provBz_data(df, round = "hour")
}
\arguments{
\item{df}{The dataframe downloaded}

\item{round}{the resampling period. Possible options "raw" "hour" "day" "week" "month"}
}
\description{
resamples data from the province of Bozen monitoring stations.
}
