% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_provBz_data2.R
\name{resample_provBz_data2}
\alias{resample_provBz_data2}
\title{resamples data from the province of Bozen monitoring stations. uses data.table}
\usage{
resample_provBz_data2(df, roundFUN, round = "hour", statsFUN, ...)
}
\arguments{
\item{df}{The data.table downloaded. Has to have colnames = "TimeStamp", "Station", "Sensor", "Value"}

\item{roundFUN}{the resampling function. Possible options lubridate::ceiling_date,lubridate::floor_date,lubridate::round_date}

\item{round}{the resampling period. Possible options "raw" "hour" "day" "week" "month". Passed to lubridate::round_date}

\item{statsFUN}{the "statistical" function. any function that takes a vector in inpunt and a single value in output}

\item{...}{additional arguments passed to statsFUN}
}
\description{
resamples data from the province of Bozen monitoring stations. uses data.table
}
