% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Get data from province of Bozen monitoring stations.}
\usage{
run_app(app, launchBrowser = T)
}
\arguments{
\item{app}{character string with the name of the app.}

\item{launchBrowser}{whether to launch the application with the browser or not.}
}
\description{
Get data from province of Bozen monitoring stations.
}
\examples{
run_app('MeteoBrowser')

}
