% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Internal function for the shiny app. Translates text into current language}
\usage{
tr(key, l, t = translation)
}
\description{
Internal function for the shiny app. Translates text into current language
}
